Fix `composite!` macro
See dogamak/xcbars#1.

Move `SubComponent` out of macro to fix trait issues. This now makes it
possible to run the examples again.
